cmd/atlas/internal/migratelint: move oss code to lint_oss
